polish all theme styles
