+++ src/pages/404.js
@@ -9,8 +9,7 @@ const NotFoundPage = () => (
     <Style.DescriptionContainer>
       <Style.ImageNotFound src={AssetJPG} alt="404-not-found" width="500" height="375"/>
       <h1>Page not Found </h1>
-      <p>We couldn&apos;t find</p>
-      <p>the page you&apos;re looking for</p>
+      <p>We couldn&apos;t find <br/> the page you&apos;re looking for</p>
       <Style.LinkCustom to="/">
         <span>Back to homepage</span>
       </Style.LinkCustom>
