+++ src/pages/theme/platinum/components/header.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from "react";
+import scrollTo from 'gatsby-plugin-smoothscroll';
 import { Grid, Box }from '@material-ui/core';
 import * as Style from "@styles/platinum/headerStyle";
 import AOS from 'aos'
@@ -16,10 +17,10 @@ const HeaderPlatinum = () =>  {
     const calculateTimeLeft = () => {
         var countDownDate = new Date("Dec 11, 2021 15:00:00").getTime();
         var now = new Date().getTime();
-    
+
         let difference = countDownDate - now;
         let timeLeft = {};
-    
+
         if (difference > 0) {
           timeLeft = {
             hari: Math.floor(difference / (1000 * 60 * 60 * 24)),
@@ -28,19 +29,19 @@ const HeaderPlatinum = () =>  {
             detik: Math.floor((difference / 1000) % 60),
           };
         }
-    
+
         return timeLeft;
     };
-    
+
     const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
       // const [year] = useState(new Date().getFullYear());
-    
+
     useEffect(() => {
     setTimeout(() => {
         setTimeLeft(calculateTimeLeft());
     }, 1000);
     });
-    
+
     const timerComponents = [];
     Object.keys(timeLeft).forEach((interval) => {
         if (!timeLeft[interval]) {
@@ -62,14 +63,14 @@ const HeaderPlatinum = () =>  {
                          <Style.Invite>Dear you, you’re invited to</Style.Invite>
                         <Style.Title >THE WEDDING OF</Style.Title>
                     </Box>
-                   
+
                     <Grid container spacing={2}>
-                        <Style.ResponsiveGrid item xs={12} md={6}>
+                        <Style.ResponsiveGrid item xs={12} md={5}>
                             <Style.Groom data-aos="fade-up">
                                 Agung
                             </Style.Groom>
                         </Style.ResponsiveGrid>
-                        <Style.ResponsiveGrid item xs={12} md={1}>
+                        <Style.ResponsiveGrid item xs={12} md={2}>
                             <Style.Ampersand data-aos="fade-up">
                                 &#38;
                             </Style.Ampersand>
@@ -83,7 +84,7 @@ const HeaderPlatinum = () =>  {
                     <Style.CountDown data-aos="fade-up" data-aos-once="true">
                         {timerComponents.length ? timerComponents : <span>11 December 2021</span>}
                     </Style.CountDown>
-                    <Style.Arrow>
+                    <Style.Arrow onClick={() => scrollTo('#intro-section')}>
                         <span></span>
                         <span></span>
                         <span></span>
