+++ src/styles/demo/platinum/headerStyle.js
@@ -26,7 +26,7 @@ export const WrapperInvitation = styled(Box)`
         top: 30px;
     }
     @media (min-width: 1024px) {
-        top: 160px;
+        top: 25%;
     }
 `;
 
@@ -51,6 +51,7 @@ export const Title = styled.p`
     line-height: 37px;
     letter-spacing: 0.3em;
     color: white;
+    margin: 35px 0px;
     @media (max-width: 767px) {
         font-size: 16px;
         line-height: 15px;
@@ -67,10 +68,12 @@ export const Groom = styled.p`
     font-weight: normal;
     font-size: 65px;
     line-height: 24px;
+    margin: 25px 0px 50px 0px;
     @media (max-width: 767px) {
         font-size: 54px;
         line-height: 12px;
         text-align: center;
+        margin: 25px 0px 25px 0px;
     }
     @media only screen and (min-width: 736px) and (max-width: 737px) {
         margin-top: 30px;
@@ -91,16 +94,19 @@ export const Ampersand = styled.p`
     font-weight: normal;
     font-size: 65px;
     line-height: 24px;
+    margin: 10px 0px;
     @media (max-width: 767px) {
         font-size: 36px;
         line-height: 36px;
         text-align: center;
-        margin: -20px 0px;
     }
     @media only screen and (min-width: 768px) and (max-width: 992px) {
         font-size: 65px;
         line-height: 24px;
         margin: -20px 0px;
+    }
+    @media (min-width: 1024px) {
+      margin: 30px 0px;
     }
     color: white;
 `;
@@ -111,6 +117,7 @@ export const Bride = styled.p`
     font-weight: normal;
     font-size: 65px;
     line-height: 24px;
+    margin: 25px 0px 50px 0px;
     @media (max-width: 767px) {
         font-size: 54px;
         line-height: 12px;
@@ -134,6 +141,9 @@ export const CountDown = styled.p`
     @media (min-width: 768px) {
         margin-top: -20px;
     }
+    @media (min-width: 1024px) {
+      margin-top: 30px;
+    }
     @media (max-width: 767px) {
         font-size: 14px;
         line-height: 17px;
@@ -142,12 +152,11 @@ export const CountDown = styled.p`
 `;
 
 export const Header = styled(Box)`
-    margin-top: -14px;
     width: 100vw;
     height: 100vh;
     background:
         linear-gradient(
-            rgba(0, 0, 0, 0.1), 
+            rgba(0, 0, 0, 0.1),
             rgba(0, 0, 0, 0.9)
         ),
         url(${AssetHeader});
@@ -156,7 +165,6 @@ export const Header = styled(Box)`
     position: relative;
     background-position: 44%;
     @media (min-width: 768px) {
-        margin-top: -16px;
         background-position: 35% 20%;
     }
     text-align: center;
